requirements.txt
# Deep Learning for 2D Materials Band Gap Prediction
# INFO5000 Course Project — HKUST(GZ)

# Deep Learning
torch>=2.0
alignn>=2024.1

# Materials Science
pymatgen>=2023.0
ase>=3.22
matminer>=0.9
jarvis-tools>=2023.11

# Data Analysis
pandas>=2.0
numpy>=1.24

# Visualization
matplotlib>=3.7
seaborn>=0.13
networkx>=3.0

# Machine Learning Baselines
scikit-learn>=1.3

# Utilities
tqdm>=4.65
